move output_directory into utils
