+++ libensemble/tests/unit_tests/test_sim_dir_properties.py
@@ -2,7 +2,7 @@ import os
 import pytest
 import shutil
 import numpy as np
-from libensemble.output_directory import EnsembleDirectory
+from libensemble.utils.output_directory import EnsembleDirectory
 from libensemble.utils.loc_stack import LocationStack
 from libensemble.utils.misc import extract_H_ranges
